+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+<!-- 2.3.2 -->
+
+### Changes in v2.3.2
+
+- Improvemants for `achievement unlocker` custom order modal
+  - Added a checkbox to each achievement
+    - Unchecked achievements will not be unlocked by `achievement unlocker`
+  - Achievements with the `hidden` property will now show a blurred description until hovered
+
 <!-- 2.3.1 -->
 
 ### Changes in v2.3.1